Split backend and extension
This patch splits the trussed-auth crate into two crates:  trussed-auth
only defines the AuthExtension and can be used by clients or other
backends implementing the extension.  trussed-auth-backend contains the
AuthBackend that implements the extension using the filesystem.
