+++ README.md
@@ -5,10 +5,14 @@ SPDX-License-Identifier: CC0-1.0
 
 # trussed-auth
 
-`trussed-auth` is an extension and custom backend for [Trussed][] that provides
-basic PIN handling.
+`trussed-auth` is an extension for [Trussed][] that provides basic PIN
+handling.  `trussed-auth-backend` is a Trussed backend implementing that
+extension using the filesystem.  Other implementations are provided by these
+backends:
+- [`trussed-se050-backend`][]
 
 [Trussed]: https://github.com/trussed-dev/trussed
+[`trussed-se050-backend`]: https://github.com/Nitrokey/trussed-se050-backend
 
 ## License
